requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
numpy==1.24.3
aiofiles==23.2.1
jinja2==3.1.2
python-multipart